requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.2
torchvision==0.17.2
transformers>=4.38.2,<5.0
pillow==10.2.0
numpy==1.26.4
pymilvus==2.4.4
opencv-python-headless==4.9.0.80
python-multipart==0.0.9
setuptools==69.5.1
marshmallow==3.20.2
paddlepaddle==2.6.2
paddleocr==2.7.0.3
rembg>=2.0.50
sentence-transformers>=2.2.0,<4.0
python-Levenshtein>=0.21.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
pyemd>=1.0.0
scipy>=1.11.0
easyocr>=1.7.0
indic-transliteration>=2.3.0
deep-translator>=1.11.0
jellyfish>=1.0.0
imagehash>=4.3.0
